Chore: 최적화잼에서 수정했던 것들 중 유용한 것들 가져옴.
	modified:   NipsEngine/NipsEngine.vcxproj
	modified:   NipsEngine/NipsEngine.vcxproj.filters
	modified:   NipsEngine/Source/Editor/Viewport/EditorViewportClient.h
	modified:   NipsEngine/Source/Editor/Viewport/ViewportCamera.h
	modified:   NipsEngine/Source/Engine/Asset/StaticMeshTypes.h
	modified:   NipsEngine/Source/Engine/Component/CameraComponent.h
	modified:   NipsEngine/Source/Engine/Component/PrimitiveComponent.cpp
	modified:   NipsEngine/Source/Engine/Component/PrimitiveComponent.h
	modified:   NipsEngine/Source/Engine/Component/SceneComponent.h
	modified:   NipsEngine/Source/Engine/Core/CoreMinimal.h
	modified:   NipsEngine/Source/Engine/Core/CrashDump.cpp
	modified:   NipsEngine/Source/Engine/Core/CrashDump.h
	modified:   NipsEngine/Source/Engine/Core/EngineTypes.h
	modified:   NipsEngine/Source/Engine/Core/Logging/GPUProfiler.cpp
	modified:   NipsEngine/Source/Engine/Core/Logging/GPUProfiler.h
	deleted:    NipsEngine/Source/Engine/Core/RayTypes.h
	new file:   NipsEngine/Source/Engine/Geometry/Edge.h
	new file:   NipsEngine/Source/Engine/Math/Edge.h
	modified:   NipsEngine/Source/Engine/Math/Matrix.h
	deleted:    NipsEngine/Source/Engine/Math/Transform.cpp
	deleted:    NipsEngine/Source/Engine/Math/Transform.h
	modified:   NipsEngine/Source/Engine/Math/Vector.h
	modified:   NipsEngine/Source/Engine/Math/Vector2.h
	new file:   NipsEngine/Source/Engine/Render/Common/ComPtr.h
	modified:   NipsEngine/Source/Engine/Render/Common/RenderTypes.h
	modified:   NipsEngine/Source/Engine/Render/Device/D3DDevice.cpp
	modified:   NipsEngine/Source/Engine/Render/Device/D3DDevice.h
	modified:   NipsEngine/Source/Engine/Render/FontBatcher.cpp
	modified:   NipsEngine/Source/Engine/Render/FontBatcher.h
	modified:   NipsEngine/Source/Engine/Render/LineBatcher.cpp
	modified:   NipsEngine/Source/Engine/Render/LineBatcher.h
	new file:   NipsEngine/Source/Engine/Render/Renderer/OcclusionCulling/OcclusionCulling.cpp
	new file:   NipsEngine/Source/Engine/Render/Renderer/OcclusionCulling/OcclusionCulling.h
	new file:   NipsEngine/Source/Engine/Render/Renderer/OcclusionCulling/Renderer.cpp
	new file:   NipsEngine/Source/Engine/Render/Renderer/OcclusionCulling/Renderer.h
	modified:   NipsEngine/Source/Engine/Render/Renderer/Renderer.cpp
	modified:   NipsEngine/Source/Engine/Render/Resource/Buffer.cpp
	modified:   NipsEngine/Source/Engine/Render/Resource/Buffer.h
	modified:   NipsEngine/Source/Engine/Render/Resource/RenderResources.h
	modified:   NipsEngine/Source/Engine/Render/Resource/Shader.cpp
	modified:   NipsEngine/Source/Engine/Render/Resource/Shader.h
	modified:   NipsEngine/Source/Engine/Render/Scene/RenderCommand.h
	modified:   NipsEngine/Source/Engine/Render/SubUVBatcher.cpp
	modified:   NipsEngine/Source/Engine/Render/SubUVBatcher.h
	modified:   NipsEngine/Source/Engine/Runtime/Launch.cpp
	modified:   NipsEngine/Source/Engine/Runtime/Launch.h
	new file:   NipsEngine/Source/Engine/Runtime/Stats/ScopeCycleCounter.h
	new file:   NipsEngine/Source/Engine/Runtime/WindowsPlatformTime.cpp
	new file:   NipsEngine/Source/Engine/Runtime/WindowsPlatformTime.h
	modified:   NipsEngine/Source/Engine/Spatial/BVH.cpp
	modified:   NipsEngine/Source/Engine/Spatial/KDTree.cpp
	modified:   NipsEngine/Source/Engine/Spatial/KDTree.h
	modified:   NipsEngine/Source/Misc/ObjViewer/UI/ObjViewerControlWidget.cpp
	modified:   NipsEngine/Source/Misc/ObjViewer/Viewport/ObjViewerViewportClient.cpp
	modified:   NipsEngine/Source/Misc/ObjViewer/Viewport/ObjViewerViewportClient.h
	modified:   NipsEngine/main.cpp

+++ NipsEngine/Source/Engine/Spatial/KDTree.cpp
@@ -18,7 +18,7 @@ void FKDTree::Build(const TArray<FNormalVertex>& Vertices, const TArray<uint32>&
 {
     Root = nullptr;
 
-    uint32 TriangleCount = Indices.size() / 3;
+	uint32 TriangleCount = static_cast<int32>(Indices.size() / 3);
     if (TriangleCount <= 0)
     {
         return;
